use a marker icon instead to show the dropdown
