+++ ui/src/app/applications/components/pod-logs-viewer/pod-logs-viewer.scss
@@ -224,72 +224,46 @@ $pod-background-dark: $dark-theme-background-2;
     }
 }
 
-.custom-select {
-    position: relative;
-    width: 300px;
+.select-container {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    background: white;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+    z-index: 1000;
+    min-width: 200px;  // Ensure dropdown is wide enough for pod names
+}
+
+.select-options {
+    max-height: 200px;
+    overflow-y: auto;
     
-    .select-header {
-        height: 37px;
-        padding: 0 10px;
-        font-size: 14px;
-        line-height: 37px;
-        cursor: pointer;
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        
-        &:after {
-            content: '';
-            position: absolute;
-            right: 10px;
-            top: 16px;
-            border-left: 5px solid transparent;
-            border-right: 5px solid transparent;
-            border-top: 5px solid #666;
-        }
-    }
-
-    .select-container {
-        position: absolute;
-        top: 100%;
-        left: 0;
-        right: 0;
-        background: white;
-        border: 1px solid #ccc;
-        border-top: none;
-        border-radius: 0 0 4px 4px;
-        box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-        z-index: 1000;
-    }
-
-    .select-options {
-        max-height: 200px;
-        overflow-y: auto;
-        border-bottom: 1px solid #ccc;
-
-        .select-option {
-            padding: 8px 10px;
-            cursor: pointer;
-            
-            &:hover {
-                background-color: #f5f5f5;
-            }
-            
-            &.selected {
-                background-color: #e6f3ff;
-            }
-        }
-    }
-
-    .clear-highlight {
+    .select-option {
         padding: 8px 10px;
-        text-align: center;
-        color: #190fd7;
         cursor: pointer;
+        white-space: nowrap;  // Prevent pod names from wrapping
         
         &:hover {
-            background-color: #f0f0f0;
+            background-color: #f5f5f5;
         }
+        
+        &.selected {
+            background-color: #e6f3ff;
+        }
+    }
+}
+
+.clear-highlight {
+    padding: 8px 10px;
+    border-top: 1px solid #ccc;
+    text-align: center;
+    color: #1750d3;
+    cursor: pointer;
+    
+    &:hover {
+        background-color: #f0f0f0;
     }
 }
